+++ Resources/plugin.yml
@@ -1,6 +1,6 @@
 main: com.sekwah.advancedportals.AdvancedPortalsPlugin
 name: AdvancedPortals
-version: 0.0.19
+version: 0.0.20
 author: sekwah41
 description: An advanced portals plugin for bukkit.
 commands:
