+++ src/com/sekwah/advancedportals/api/PortalArg.java
@@ -1,4 +1,4 @@
-package com.sekwah.advancedportals.portals;
+package com.sekwah.advancedportals.api;
 
 public class PortalArg {
